+++ NEWS.md
@@ -1,9 +1,9 @@
 PvSTATEM 0.0.3
 ---------------------------------------------------------------
-* refactored whole package structure to simplify the usage
+* refactored the whole package structure to simplify the usage
 * new, much faster parser for the xPONENT and INTELLIFLEX files
 * model encapsulation for the standard curve fitting
-* new plots - mfi chart and layout plot
+* new plots - MFI chart and layout plot
 * updated blank adjustment function
 * error fixes
 
@@ -12,7 +12,7 @@ PvSTATEM 0.0.2
 ---------------------------------------------------------------
 * added new datasets to the package
 * sample standard curve plotting
-* github actions
+* GitHub actions
 * tests with coverage calculation